Disable terminal coloring for all filetypes except text
probably does not make sense to try to color those lines in files other
than of type text or when the filetype is not set (logfiles). Since this
is kind of costly, disable it.

Also, while at it, move the check for the disabled feature up a bit, so
Vim does not first check for a timeout just to notice afterwards that it
has nothing to do here.

Should also fix #68